requirements.txt
google-genai
supabase
langchain-text-splitters
tiktoken
python-dotenv
fastapi
uvicorn
